feat: implement Confirm dialog

+++ project-management-app/project-management-app/src/app/pages/components/mainpage/components/boards-list/boards-list.component.ts
@@ -2,6 +2,7 @@ import { Component,  OnInit, OnDestroy, Output } from '@angular/core';
 import { Observable, Subscription } from 'rxjs';
 import { Keys } from 'src/app/shared/models/enums/key-enum';
 import { Board, CreateBoardEvent } from 'src/app/shared/models/interfaces/board-interface';
+import { DialogService } from 'src/app/shared/services/dialog/dialog.service';
 import { MainpageService } from 'src/app/shared/services/mainpage/mainpage.service';
 import { StorageService } from 'src/app/shared/services/storage/storage.service';
 
@@ -14,12 +15,13 @@ export class BoardsListComponent implements OnInit, OnDestroy {
   boards$: Observable<Board[]> = this.mainpageService.getAllBoards$();
   boards: Board[] = [];
   private subscription: Subscription[] = [];
-  public searchText: string;
+  searchText: string;
 
 
   constructor(
     private mainpageService: MainpageService,
-    private storageSernice: StorageService) {}
+    private storageSernice: StorageService,
+    private dialogService: DialogService) {}
 
   ngOnInit() {
     this.mainpageService.getAllBoard();
@@ -40,8 +42,18 @@ export class BoardsListComponent implements OnInit, OnDestroy {
     }
   }
 
-  public deleteBoard(id: string) {
+  deleteBoard(eventConfirm: any, id: string) {
+    if (eventConfirm) {
       this.mainpageService.deleteBoard(id);
+    }
+  }
+
+  setOneField() {
+    this.dialogService.setOneFiled();
+  }
+
+  setDeleteBoard() {
+    this.dialogService.setDeleteBoard();
   }
 
   ngOnDestroy() {
